+++ packages/storybook/config/storybook/preview.js
@@ -1,6 +1,6 @@
 import Vue from 'vue';
 
 // Setup Context as a Web Host
-import '../../../../test/context/storybook.entry';
+import '../../../../services/f-development-context/storybook.entry';
 
 Vue.config.devtools = true;
